+++ smtlib-backends-process/src/SMTLIB/Backends/Process.hs
@@ -16,9 +16,7 @@ module SMTLIB.Backends.Process
   )
 where
 
-import Control.Concurrent.Async (Async, async, cancel)
 import qualified Control.Exception as X
-import Control.Monad (forever)
 import Data.ByteString.Builder
   ( Builder,
     byteString,
@@ -26,7 +24,6 @@ import Data.ByteString.Builder
     toLazyByteString,
   )
 import qualified Data.ByteString.Char8 as BS
-import qualified Data.ByteString.Lazy.Char8 as LBS
 import Data.Default (Default, def)
 import GHC.IO.Exception (IOException (ioe_description))
 import SMTLIB.Backends (Backend (..))
@@ -37,19 +34,14 @@ data Config = Config
   { -- | The command to call to run the solver.
     exe :: String,
     -- | Arguments to pass to the solver's command.
-    args :: [String],
-    -- | A function for logging the solver process' messages on stderr and file
-    -- handle exceptions.
-    -- If you want line breaks between each log message, you need to implement
-    -- it yourself, e.g use @'LBS.putStr' . (<> "\n")@.
-    reportError :: LBS.ByteString -> IO ()
+    args :: [String]
   }
 
 -- | By default, use Z3 as an external process and ignores log messages.
 defaultConfig :: Config
 -- if you change this, make sure to also update the comment two lines above
 -- as well as the one in @smtlib-backends-process/tests/Examples.hs@
-defaultConfig = Config "z3" ["-in"] (\_ -> return ())
+defaultConfig = Config "z3" ["-in"]
 
 instance Default Config where
   def = defaultConfig
@@ -62,9 +54,7 @@ data Handle = Handle
     -- | The output channel of the process.
     hOut :: IO.Handle,
     -- | The error channel of the process.
-    hErr :: IO.Handle,
-    -- | A process reading the solver's error messages and logging them.
-    errorReader :: Async ()
+    hErr :: IO.Handle
   }
 
 -- | Run a solver as a process.
@@ -82,21 +72,11 @@ new Config {..} = decorateIOError "creating the solver process" $ do
         }
   mapM_ setupHandle [hIn, hOut, hErr]
   -- log error messages created by the backend
-  errorReader <-
-    async $
-      forever
-        ( do
-            errs <- BS.hGetLine hErr
-            reportError' errs
-        )
-        `X.catch` \X.SomeException {} ->
-          return ()
-  return $ Handle process hIn hOut hErr errorReader
+  return $ Handle process hIn hOut hErr
   where
     setupHandle h = do
       IO.hSetBinaryMode h True
       IO.hSetBuffering h $ IO.BlockBuffering Nothing
-    reportError' = reportError . LBS.fromStrict
 
 -- | Send a command to the process without reading its response.
 write :: Handle -> Builder -> IO ()
@@ -110,7 +90,6 @@ write Handle {..} cmd =
 -- | Cleanup the process' resources, terminate it and wait for it to actually exit.
 close :: Handle -> IO ()
 close Handle {..} = decorateIOError "closing the solver process" $ do
-  cancel errorReader
   P.cleanupProcess (Just hIn, Just hOut, Just hErr, process)
 
 -- | Create a solver process, use it to make a computation and close it.